CPoints contains longs, not ints
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@1533 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/Gme/GMEView.h
@@ -118,10 +118,10 @@ public:
 	CPoint					customizeConnectionEdgeStartPoint;
 	CPoint					customizeConnectionEdgeEndPoint;
 	CPoint					customizeConnectionEdgeThirdPoint;
-	int						customizeConnectionEdgeXMinLimit;
-	int						customizeConnectionEdgeXMaxLimit;
-	int						customizeConnectionEdgeYMinLimit;
-	int						customizeConnectionEdgeYMaxLimit;
+	long						customizeConnectionEdgeXMinLimit;
+	long						customizeConnectionEdgeXMaxLimit;
+	long						customizeConnectionEdgeYMinLimit;
+	long						customizeConnectionEdgeYMaxLimit;
 	CPoint					customizeConnectionOrigCursor;
 	CPoint					customizeConnectionCurrCursor;
 	bool					isCursorChangedByEdgeCustomize;
